+++ src/components/header/header_links/header_links.test.tsx
@@ -21,12 +21,45 @@ import React from 'react';
 import { render } from 'enzyme';
 import { requiredProps } from '../../../test';
 
-import { EuiHeaderLinks } from './header_links';
+import { EuiHeaderLinks, GUTTER_SIZES } from './header_links';
 
 describe('EuiHeaderLinks', () => {
   test('is rendered', () => {
     const component = render(<EuiHeaderLinks {...requiredProps} />);
 
     expect(component).toMatchSnapshot();
+  });
+
+  describe('gutterSize', () => {
+    GUTTER_SIZES.forEach(gutterSize => {
+      test(`${gutterSize} is rendered`, () => {
+        const component = render(<EuiHeaderLinks gutterSize={gutterSize} />);
+
+        expect(component).toMatchSnapshot();
+      });
+    });
+  });
+
+  describe('popover props', () => {
+    test('is rendered', () => {
+      const component = render(
+        <EuiHeaderLinks
+          popoverBreakpoints={['xs', 's', 'm', 'l', 'xl']}
+          popoverButtonProps={{
+            iconType: 'bolt',
+            className: 'customButtonClass',
+          }}
+          popoverProps={{ anchorClassName: 'customAnchorClass' }}
+        />
+      );
+
+      expect(component).toMatchSnapshot();
+    });
+
+    test('is never rendered with "none"', () => {
+      const component = render(<EuiHeaderLinks popoverBreakpoints={'none'} />);
+
+      expect(component).toMatchSnapshot();
+    });
   });
 });